+++ examples/ibm-private-dns/main.tf
@@ -196,9 +196,24 @@ resource "ibm_dns_custom_resolver" "test" {
     enabled     = false
   }
 }
+resource "ibm_dns_custom_resolver" "test" {
+  name            = "testCR-TF-New"
+  instance_id     = ibm_resource_instance.test-pdns-instance.guid
+  description     = "new test CR TF-1"
+  high_availability = true
+  enabled       = true
+  locations {
+    subnet_crn = "crn:v1:staging:public:is:us-south-1:a/01652b251c3ae2787110a995d8db0135::subnet:0716-a094c4e8-02cd-4b04-858d-3432"
+    enabled     = false
+  } 
+  locations {
+    subnet_crn  = "crn:v1:staging:public:is:us-south-2:a/01652b251c3ae2787110a995d8db0135::subnet:0726-b6f3cb83-48f0-4c55-9023-433"
+    enabled     = false
+  }
+}
 
-data "ibm_dns_custom_resolvers" "test-custom-resolver" {
-  instance_id = ibm_dns_custom_resolver.test.instance_id
+data "ibm_dns_custom_resolvers" "test-cr" {
+		instance_id = ibm_dns_custom_resolver.test.instance_id
 }
 
 output "ibm_dns_custom_resolvers_output" {
@@ -206,22 +221,23 @@ output "ibm_dns_custom_resolvers_output" {
 }
 
 resource "ibm_dns_custom_resolver_location" "test" {
+  instance_id   = ibm_resource_instance.test-pdns-instance.guid
+  resolver_id   = ibm_dns_custom_resolver.test.custom_resolver_id
+  subnet_crn    = "crn:v1:staging:public:is:us-south-1:a/01652b251c3ae2787110a995d8db0135::subnet:0716-a094c4e8-02cd-4b04-858d-343"
+  enabled       = true
+  cr_enabled    = true
+} 
+
+resource "ibm_dns_custom_resolver_forwarding_rule" "test" {
   instance_id = ibm_resource_instance.test-pdns-instance.guid
   resolver_id = ibm_dns_custom_resolver.test.custom_resolver_id
-  subnet_crn  = "crn:v1:staging:public:is:us-south-1:a/01652b251c3ae2787110a995d8db0135::subnet:0716-03d54d71-b438-4d20-b943-76d3d2a1a590"
-  enabled     = false
-}
-
-resource "ibm_dns_custom_resolver_forwarding_rule" "dns_custom_resolver_forwarding_rule" {
- instance_id = ibm_resource_instance.test-pdns-instance.guid
- resolver_id = ibm_dns_custom_resolver.test.custom_resolver_id
   description = "test forward rule"
   type = "zone"
   match = "test.example.com"
   forward_to = ["168.20.22.122"]
 }
 
-data "ibm_dns_custom_resolver_forwarding_rules" "dns_custom_resolver_forwarding_rules" {
-  instance_id = ibm_dns_custom_resolver_forwarding_rule.dns_custom_resolver_forwarding_rule.instance_id
-  resolver_id = ibm_dns_custom_resolver_forwarding_rule.dns_custom_resolver_forwarding_rule.resolver_id
+data "ibm_dns_custom_resolver_forwarding_rules" "test-fr" {
+		instance_id	= ibm_dns_custom_resolver.test.instance_id
+		resolver_id = ibm_dns_custom_resolver.test.custom_resolver_id
 }