+++ ibm/data_source_ibm_dl_gateways.go
@@ -36,6 +36,26 @@ func dataSourceIBMDLGateways() *schema.Resource {
 							Computed:    true,
 							Description: "BGP MD5 authentication key",
 						},
+						dlBfdInterval: {
+							Type:        schema.TypeInt,
+							Computed:    true,
+							Description: "BFD Interval",
+						},
+						dlBfdMultiplier: {
+							Type:        schema.TypeInt,
+							Computed:    true,
+							Description: "BFD Multiplier",
+						},
+						dlBfdStatus: {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "BFD Status",
+						},
+						dlBfdStatusUpdatedAt: {
+							Type:        schema.TypeString,
+							Computed:    true,
+							Description: "BFD Status",
+						},
 						dlBgpAsn: {
 							Type:        schema.TypeInt,
 							Computed:    true,
@@ -322,6 +342,26 @@ func dataSourceIBMDLGatewaysRead(d *schema.ResourceData, meta interface{}) error
 			rg := instance.ResourceGroup
 			gateway[dlResourceGroup] = *rg.ID
 		}
+
+		//Show the BFD Config parameters if set
+		if instance.BfdConfig != nil {
+			if instance.BfdConfig.Interval != nil {
+				gateway[dlBfdInterval] = *instance.BfdConfig.Interval
+			}
+
+			if instance.BfdConfig.Multiplier != nil {
+				gateway[dlBfdMultiplier] = *instance.BfdConfig.Multiplier
+			}
+
+			if instance.BfdConfig.BfdStatus != nil {
+				gateway[dlBfdStatus] = *instance.BfdConfig.BfdStatus
+			}
+
+			if instance.BfdConfig.BfdStatusUpdatedAt != nil {
+				gateway[dlBfdStatusUpdatedAt] = instance.BfdConfig.BfdStatusUpdatedAt.String()
+			}
+		}
+
 		dtype := *instance.Type
 		if dtype == "dedicated" {
 			if instance.MacsecConfig != nil {