+++ ibm/resource_ibm_is_instance.go
@@ -39,6 +39,9 @@ const (
 	isInstanceBootVolume              = "boot_volume"
 	isInstanceVolumeSnapshot          = "snapshot"
 	isInstanceSourceTemplate          = "instance_template"
+	isInstanceBandwidth               = "bandwidth"
+	isInstanceTotalVolumeBandwidth    = "total_volume_bandwidth"
+	isInstanceTotalNetworkBandwidth   = "total_network_bandwidth"
 	isInstanceVolAttVolAutoDelete     = "auto_delete_volume"
 	isInstanceVolAttVolBillingTerm    = "billing_term"
 	isInstanceImage                   = "image"
@@ -67,6 +70,8 @@ const (
 	isInstanceDeleteDone              = "done"
 	isInstanceFailed                  = "failed"
 
+	isInstanceStatusRestarting     = "restarting"
+	isInstanceStatusStarting       = "starting"
 	isInstanceActionStatusStopping = "stopping"
 	isInstanceActionStatusStopped  = "stopped"
 	isInstanceStatusPending        = "pending"
@@ -78,12 +83,12 @@ const (
 	isInstanceBootIOPS           = "iops"
 	isInstanceBootEncryption     = "encryption"
 	isInstanceBootProfile        = "profile"
-
-	isInstanceVolumeAttachments = "volume_attachments"
-	isInstanceVolumeAttaching   = "attaching"
-	isInstanceVolumeAttached    = "attached"
-	isInstanceVolumeDetaching   = "detaching"
-	isInstanceResourceGroup     = "resource_group"
+	isInstanceAction             = "action"
+	isInstanceVolumeAttachments  = "volume_attachments"
+	isInstanceVolumeAttaching    = "attaching"
+	isInstanceVolumeAttached     = "attached"
+	isInstanceVolumeDetaching    = "detaching"
+	isInstanceResourceGroup      = "resource_group"
 
 	isPlacementTargetDedicatedHost      = "dedicated_host"
 	isPlacementTargetDedicatedHostGroup = "dedicated_host_group"
@@ -212,6 +217,26 @@ func resourceIBMISInstance() *schema.Resource {
 				Description:   "Unique Identifier of the Placement Group for restricting the placement of the instance",
 			},
 
+			isInstanceTotalVolumeBandwidth: {
+				Type:         schema.TypeInt,
+				Optional:     true,
+				Computed:     true,
+				ValidateFunc: InvokeValidator("ibm_is_instance", isInstanceTotalVolumeBandwidth),
+				Description:  "The amount of bandwidth (in megabits per second) allocated exclusively to instance storage volumes",
+			},
+
+			isInstanceBandwidth: {
+				Type:        schema.TypeInt,
+				Computed:    true,
+				Description: "The total bandwidth (in megabits per second) shared across the instance's network interfaces and storage volumes",
+			},
+
+			isInstanceTotalNetworkBandwidth: {
+				Type:        schema.TypeInt,
+				Computed:    true,
+				Description: "The amount of bandwidth (in megabits per second) allocated exclusively to instance network interfaces.",
+			},
+
 			isInstanceKeys: {
 				Type:             schema.TypeSet,
 				Optional:         true,
@@ -236,6 +261,21 @@ func resourceIBMISInstance() *schema.Resource {
 				Default:          true,
 				DiffSuppressFunc: suppressEnableCleanDelete,
 				Description:      "Enables stopping of instance before deleting and waits till deletion is complete",
+			},
+
+			isInstanceAction: {
+				Type:         schema.TypeString,
+				Optional:     true,
+				ValidateFunc: InvokeValidator("ibm_is_instance", isInstanceAction),
+				Description:  "Enables stopping of instance before deleting and waits till deletion is complete",
+			},
+
+			isInstanceActionForce: {
+				Type:         schema.TypeBool,
+				Optional:     true,
+				RequiredWith: []string{isInstanceAction},
+				Default:      false,
+				Description:  "If set to true, the action will be forced immediately, and all queued actions deleted. Ignored for the start action.",
 			},
 
 			isInstanceVolumeAttachments: {
@@ -362,11 +402,10 @@ func resourceIBMISInstance() *schema.Resource {
 			},
 
 			isInstanceUserData: {
-				Type:             schema.TypeString,
-				ForceNew:         true,
-				DiffSuppressFunc: suppressUserData,
-				Optional:         true,
-				Description:      "User data given for the instance",
+				Type:        schema.TypeString,
+				ForceNew:    true,
+				Optional:    true,
+				Description: "User data given for the instance",
 			},
 
 			isInstanceImage: {
@@ -656,7 +695,7 @@ func resourceIBMISInstance() *schema.Resource {
 }
 
 func resourceIBMISInstanceValidator() *ResourceValidator {
-
+	actions := "stop, start, reboot"
 	validateSchema := make([]ValidateSchema, 0)
 	validateSchema = append(validateSchema,
 		ValidateSchema{
@@ -676,6 +715,20 @@ func resourceIBMISInstanceValidator() *ResourceValidator {
 			Regexp:                     `^[A-Za-z0-9:_ .-]+$`,
 			MinValueLength:             1,
 			MaxValueLength:             128})
+	validateSchema = append(validateSchema,
+		ValidateSchema{
+			Identifier:                 isInstanceTotalVolumeBandwidth,
+			ValidateFunctionIdentifier: IntAtLeast,
+			Type:                       TypeInt,
+			Optional:                   true,
+			MinValue:                   "500"})
+	validateSchema = append(validateSchema,
+		ValidateSchema{
+			Identifier:                 isInstanceAction,
+			ValidateFunctionIdentifier: ValidateAllowedStringValue,
+			Type:                       TypeString,
+			Optional:                   true,
+			AllowedValues:              actions})
 
 	validateSchema = append(validateSchema,
 		ValidateSchema{
@@ -711,7 +764,10 @@ func instanceCreateByImage(d *schema.ResourceData, meta interface{}, profile, na
 			ID: &vpcID,
 		},
 	}
-
+	if totalVolBandwidthIntf, ok := d.GetOk(isInstanceTotalVolumeBandwidth); ok {
+		totalVolBandwidthStr := int64(totalVolBandwidthIntf.(int))
+		instanceproto.TotalVolumeBandwidth = &totalVolBandwidthStr
+	}
 	if dHostIdInf, ok := d.GetOk(isPlacementTargetDedicatedHost); ok {
 		dHostIdStr := dHostIdInf.(string)
 		dHostPlaementTarget := &vpcv1.InstancePlacementTargetPrototypeDedicatedHostIdentity{
@@ -918,6 +974,11 @@ func instanceCreateByTemplate(d *schema.ResourceData, meta interface{}, profile,
 			Name: &profile,
 		}
 	}
+	if totalVolBandwidthIntf, ok := d.GetOk(isInstanceTotalVolumeBandwidth); ok {
+		totalVolBandwidthStr := int64(totalVolBandwidthIntf.(int))
+		instanceproto.TotalVolumeBandwidth = &totalVolBandwidthStr
+	}
+
 	if vpcID != "" {
 		instanceproto.VPC = &vpcv1.VPCIdentity{
 			ID: &vpcID,
@@ -1189,6 +1250,10 @@ func instanceCreateByVolume(d *schema.ResourceData, meta interface{}, profile, n
 			DeleteVolumeOnInstanceDelete: &deletebool,
 			Volume:                       volTemplate,
 		}
+	}
+	if totalVolBandwidthIntf, ok := d.GetOk(isInstanceTotalVolumeBandwidth); ok {
+		totalVolBandwidthStr := int64(totalVolBandwidthIntf.(int))
+		instanceproto.TotalVolumeBandwidth = &totalVolBandwidthStr
 	}
 
 	if primnicintf, ok := d.GetOk(isInstancePrimaryNetworkInterface); ok {
@@ -1495,6 +1560,18 @@ func instanceGet(d *schema.ResourceData, meta interface{}, id string) error {
 	}
 	d.Set(isInstanceCPU, cpuList)
 
+	if instance.Bandwidth != nil {
+		d.Set(isInstanceBandwidth, int(*instance.Bandwidth))
+	}
+
+	if instance.TotalNetworkBandwidth != nil {
+		d.Set(isInstanceTotalNetworkBandwidth, int(*instance.TotalNetworkBandwidth))
+	}
+
+	if instance.TotalVolumeBandwidth != nil {
+		d.Set(isInstanceTotalVolumeBandwidth, int(*instance.TotalVolumeBandwidth))
+	}
+
 	d.Set(isInstanceMemory, *instance.Memory)
 	gpuList := make([]map[string]interface{}, 0)
 	if instance.Gpu != nil {
@@ -1680,6 +1757,51 @@ func instanceUpdate(d *schema.ResourceData, meta interface{}) error {
 		return err
 	}
 	id := d.Id()
+
+	if d.HasChange(isInstanceAction) && !d.IsNewResource() {
+
+		actiontype := d.Get(isInstanceAction).(string)
+		if actiontype != "" {
+			getinsOptions := &vpcv1.GetInstanceOptions{
+				ID: &id,
+			}
+			instance, response, err := instanceC.GetInstance(getinsOptions)
+			if err != nil {
+				return fmt.Errorf("Error Getting Instance (%s): %s\n%s", id, err, response)
+			}
+			if (actiontype == "stop" || actiontype == "reboot") && *instance.Status != isInstanceStatusRunning {
+				d.Set(isInstanceAction, nil)
+				return fmt.Errorf("Error with stop/reboot action: Cannot invoke stop/reboot action while instance is not in running state")
+			} else if actiontype == "start" && *instance.Status != isInstanceActionStatusStopped {
+				d.Set(isInstanceAction, nil)
+				return fmt.Errorf("Error with start action: Cannot invoke start action while instance is not in stopped state")
+			}
+			createinsactoptions := &vpcv1.CreateInstanceActionOptions{
+				InstanceID: &id,
+				Type:       &actiontype,
+			}
+			if instanceActionForceIntf, ok := d.GetOk(isInstanceActionForce); ok {
+				force := instanceActionForceIntf.(bool)
+				createinsactoptions.Force = &force
+			}
+			_, response, err = instanceC.CreateInstanceAction(createinsactoptions)
+			if err != nil {
+				return fmt.Errorf("Error Creating Instance Action: %s\n%s", err, response)
+			}
+			if actiontype == "stop" {
+				_, err = isWaitForInstanceActionStop(instanceC, d.Timeout(schema.TimeoutUpdate), id, d)
+				if err != nil {
+					return err
+				}
+			} else if actiontype == "start" || actiontype == "reboot" {
+				_, err = isWaitForInstanceActionStart(instanceC, d.Timeout(schema.TimeoutUpdate), id, d)
+				if err != nil {
+					return err
+				}
+			}
+
+		}
+	}
 	if d.HasChange(isInstanceVolumes) {
 		ovs, nvs := d.GetChange(isInstanceVolumes)
 		ov := ovs.(*schema.Set)
@@ -1896,6 +2018,27 @@ func instanceUpdate(d *schema.ResourceData, meta interface{}) error {
 			}
 		}
 
+	}
+
+	if d.HasChange(isInstanceTotalVolumeBandwidth) && !d.IsNewResource() {
+		totalVolBandwidth := int64(d.Get(isInstanceTotalVolumeBandwidth).(int))
+		updnetoptions := &vpcv1.UpdateInstanceOptions{
+			ID: &id,
+		}
+
+		instancePatchModel := &vpcv1.InstancePatch{
+			TotalVolumeBandwidth: &totalVolBandwidth,
+		}
+		instancePatch, err := instancePatchModel.AsPatch()
+		if err != nil {
+			return fmt.Errorf("Error calling asPatch with total volume bandwidth for InstancePatch: %s", err)
+		}
+		updnetoptions.InstancePatch = instancePatch
+
+		_, _, err = instanceC.UpdateInstance(updnetoptions)
+		if err != nil {
+			return err
+		}
 	}
 
 	if d.HasChange(isInstanceName) && !d.IsNewResource() {
@@ -2215,6 +2358,45 @@ func isWaitForInstanceActionStop(instanceC *vpcv1.VpcV1, timeout time.Duration, 
 	return stateConf.WaitForState()
 }
 
+func isWaitForInstanceActionStart(instanceC *vpcv1.VpcV1, timeout time.Duration, id string, d *schema.ResourceData) (interface{}, error) {
+	communicator := make(chan interface{})
+	stateConf := &resource.StateChangeConf{
+		Pending: []string{isInstanceActionStatusStopped, isInstanceStatusPending, isInstanceActionStatusStopping, isInstanceStatusStarting, isInstanceStatusRestarting},
+		Target:  []string{isInstanceStatusRunning, isInstanceStatusFailed, ""},
+		Refresh: func() (interface{}, string, error) {
+			getinsoptions := &vpcv1.GetInstanceOptions{
+				ID: &id,
+			}
+			instance, response, err := instanceC.GetInstance(getinsoptions)
+			if err != nil {
+				return nil, "", fmt.Errorf("Error Getting Instance: %s\n%s", err, response)
+			}
+			select {
+			case data := <-communicator:
+				return nil, "", data.(error)
+			default:
+				fmt.Println("no message sent")
+			}
+			if *instance.Status == isInstanceStatusFailed {
+				// let know the isRestartStopAction() to stop
+				close(communicator)
+				return instance, *instance.Status, fmt.Errorf("The  instance %s failed to start: %v", id, err)
+			}
+			return instance, *instance.Status, nil
+		},
+		Timeout:    timeout,
+		Delay:      10 * time.Second,
+		MinTimeout: 10 * time.Second,
+	}
+
+	if v, ok := d.GetOk("force_recovery_time"); ok {
+		forceTimeout := v.(int)
+		go isRestartStopAction(instanceC, id, d, forceTimeout, communicator)
+	}
+
+	return stateConf.WaitForState()
+}
+
 func isRestartStopAction(instanceC *vpcv1.VpcV1, id string, d *schema.ResourceData, forceTimeout int, communicator chan interface{}) {
 	subticker := time.NewTicker(time.Duration(forceTimeout) * time.Minute)
 	//subticker := time.NewTicker(time.Duration(forceTimeout) * time.Second)
@@ -2321,14 +2503,6 @@ func resourceIbmIsInstanceInstanceDiskToMap(instanceDisk vpcv1.InstanceDisk) map
 }
 
 func suppressEnableCleanDelete(k, old, new string, d *schema.ResourceData) bool {
-	// During import
-	if old == "" && d.Id() != "" {
-		return true
-	}
-	return false
-}
-
-func suppressUserData(k, old, new string, d *schema.ResourceData) bool {
 	// During import
 	if old == "" && d.Id() != "" {
 		return true